Improves model point calculation with median method
Refactors the model point calculation to allow selecting points
closest to either the mean or the median of the values. It
introduces a `method` parameter to choose between the two. It
also filters the values to only consider those less than or
equal to the target value (mean or median) to avoid outliers.
This makes the method more adaptable and robust.